fix: in case of unsupported platform - install dcontainer as python pipx package

+++ dcontainer/cli/generate/file_models/dependencies_sh.py
@@ -20,8 +20,42 @@ HEADER = """#!/usr/bin/env bash
 # For more information: https://github.com/devcontainers-contrib/cli 
 
 set -e
+ensure_jq() {{
+    # Ensure JQ available
+    if ! type jq >/dev/null 2>&1; then
+        apt-get update -y && apt-get -y install --no-install-recommends jq
+    fi 
+}}
 
 
+ensure_pipx() {{
+    # Ensure the existance of minimal python3 and pipx
+    # If no python - install it
+	if ! type python3 >/dev/null 2>&1; then
+		echo "installing python3-minimal libffi-dev"
+		apt-get update -y
+		apt-get -y install python3-minimal
+	fi
+
+	# If no pip - install it
+	if ! type pip3 >/dev/null 2>&1; then
+		echo "installing python3-pip"
+		apt-get update -y
+		apt-get -y install libffi-dev python3-pip
+	fi
+
+    # If ensurepip fails - install python venv
+	if ! python3 -Im ensurepip --version >/dev/null 2>&1; then
+		echo "installing python3-venv"
+		apt-get update -y
+		apt-get -y install python3-venv
+	fi
+
+    # If no pipx - install it
+    if ! pipx >/dev/null 2>&1; then
+        pip3 install pipx
+    fi
+}}
 
 ensure_curl() {{
     # Ensure curl available
@@ -49,21 +83,31 @@ ensure_dcontainer() {{
 
     # If not previuse installation found, download it temporarly and delete at the end of the script 
     if [[ -z "${{dcontainer_location}}" ]]; then
-        tmp_dir=$(mktemp -d -t dcontainer-XXXXXXXXXX)
 
-        clean_up () {{
-            ARG=$?
-            rm -rf $tmp_dir
-            exit $ARG
-        }}
+        if [ "$(uname -sm)" != "Linux x86_64" ]; then
+            # No binaries compiled for non linux x*^ yet, therefor we fallback to install through python
 
-        trap clean_up EXIT
+            ensure_pipx
+            ensure_jq
+            pipx install dcontainer=={RELEASE_VERSION}
+            dcontainer_location=$(pipx list --json | jq ".venvs.dcontainer.metadata.main_package.app_paths[0].__Path__" | tr -d '"')
 
-        curl -sSL -o $tmp_dir/dcontainer {dcontainer_link} 
-        curl -sSL -o $tmp_dir/checksums.txt {checksums_link}
-        (cd $tmp_dir ; sha256sum --check --strict --ignore-missing $tmp_dir/checksums.txt)
-        chmod a+x $tmp_dir/dcontainer
-        dcontainer_location=$tmp_dir/dcontainer
+        else
+            tmp_dir=$(mktemp -d -t dcontainer-XXXXXXXXXX)
+
+            clean_up () {{
+                ARG=$?
+                rm -rf $tmp_dir
+                exit $ARG
+            }}
+            trap clean_up EXIT
+
+            curl -sSL -o $tmp_dir/dcontainer {dcontainer_link} 
+            curl -sSL -o $tmp_dir/checksums.txt {checksums_link}
+            (cd $tmp_dir ; sha256sum --check --strict --ignore-missing $tmp_dir/checksums.txt)
+            chmod a+x $tmp_dir/dcontainer
+            dcontainer_location=$tmp_dir/dcontainer
+        fi
     fi
 
     # Expose outside the resolved location
@@ -171,4 +215,5 @@ class DependenciesSH(File):
             checksums_link=CHECKSUM_LINK.format(RELEASE_VERSION=self.release_version),
             force_cli_installation_env=ENV_FORCE_CLI_INSTALLATION,
             cli_location_env=ENV_CLI_LOCATION,
+            RELEASE_VERSION=self.release_version
         )
